InicioPython/requirements-psycopg2.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
marshmallow==3.21.1
bcrypt==4.1.2
flask-jwt-extended==4.6.0




